Extract StateManager from Main window
